requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
Django==5.1.5
django-admin-interface==0.29.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-colorfield==0.11.0
django-jazzmin==3.0.1
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
eventlet==0.38.2
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
msgpack==1.1.0
pillow==11.1.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
redis==5.2.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
